add triangle export to wasm

+++ geometry/src/shapes/triangle.rs
@@ -6,7 +6,7 @@ use crate::str_to_i32;
 pub struct Triangle {
     pub name: String,
     pub points: Points,
-    pub indices: Vec<i8>,
+    pub indices: Vec<i32>,
     pub origin: Point,
     pub aabb: BoundingBox
 }
